problem with node11.11.0 and jest called for jest update. Also used the commit requirement to restructure project and dump the source into functional holding directories
