+++ package.json
@@ -24,7 +24,7 @@
     "request": "^2.88.0"
   },
   "devDependencies": {
-    "jest": "^24.1.0",
+    "jest": "^24.3.0",
     "nodemon": "^1.18.10"
   }
 }
