+++ src/blockchain/block.js
@@ -9,8 +9,8 @@
  */
 
 const hexToBinary = require('hex-to-binary')
-const { GENESIS_DATA, MINE_RATE } = require('../config/config')
-const cryptoHash = require('./crypto-hash')
+const { GENESIS_DATA, MINE_RATE } = require('../../config/config')
+const cryptoHash = require('../helpers/crypto-hash')
 
 class Block {
   constructor({ timestamp, lastHash, hash, data, nonce, difficulty }) {
