+++ src/blockchain/blockchain.js
@@ -1,5 +1,5 @@
 const Block = require('./block')
-const cryptoHash = require('./crypto-hash')
+const cryptoHash = require('../helpers/crypto-hash')
 
 class Blockchain {
   constructor() {
