+++ src/helpers/crypto-hash.js
